Moved nsbundle tests to a different folder as they fail SPM test
